+++ src/main/java/net/ttddyy/dsproxy/listener/logging/CommonsQueryLoggingListener.java
@@ -23,9 +23,23 @@ public class CommonsQueryLoggingListener extends AbstractQueryLoggingListener {
         this.logLevel = logLevel;
     }
 
+    /**
+     * @deprecated use {{@link #setLog(String)}}
+     */
     @Override
+    @Deprecated
     protected void resetLogger(String loggerName) {
         this.log = LogFactory.getLog(loggerName);
+    }
+
+    /**
+     * Override {@link Log} instance with specified log name.
+     *
+     * @param logName log name
+     * @since 1.4.1
+     */
+    public void setLog(String logName) {
+        this.log = LogFactory.getLog(logName);
     }
 
     /**